Move multer setup to middleware directory
- Separated multer configuration from cloudinary module
- Created a new middleware folder to house multer setup
